【WA】[AtCoder Regular Contest 008]C - THE☆たこ焼き祭り2012

+++ arc008/arc008_3.cpp
@@ -128,7 +128,7 @@ int main() {
             long double xx = x * x;
             long double yy = y * y;
 
-            long double dis = sqrt(xx + yy);
+            long double dis = sqrtl(xx + yy);
             long double time = dis / speed;
             dijkstra.insert(i, j, time);
